[op][n_upsable][low_end_mode][webos] Reduce discardable memory limits
:Release Notes:
Better default for discardable memory in low endwebOS, to avoid
expunging background applications from memory manager when not needed.

:Detailed Notes:
We are using in webOS the default settings for discardable memory
designed for desktop browsers. We keep it for default webOS to avoid
changing the default behavior.

But, on low end modes, we change it to be quite aggressive, 16MB.
Hopefully enough anyway in the contrained case.

In low end devices, we set a lower discardable memory limit
automatically. Add a runtime features to allow to force enabling or
disabling it from command line:
- LowEndDiscardableMemoryLimit (enabled by default, sets the low end
limit for low end devices). If explicitely set, it will also apply
on non low end devices.

On top of that, we add a new feature, DiscardableMemoryCustomLimit,
with a parameter, limit-mb, to tune an specific value, and extra
logging. I.e. to set 32MB:
  --enable-feature=DiscardableMemoryCustomLimit:limit-mb/32

:Testing Performed:
1. Launch several applications.
2. Play a video in one of them.

:QA Notes:
Test-cases: QWTC-14542

:Issues Addressed:
[WRQ-3650] Support base::SysInfo::AmountOfPhysicalMemoryMB() and IsLowEndDevice
[WRQ-26199] Make a TC and integrate low end device mode patches to chromium120

Change-Id: I0d1c0f614cc0373557ecdd60f64741f91eee3cbf
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/414186
Blink-Build: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/419090
Tested-by: Blink Jenkins <blink.jenkins@lge.com>
Blink-Build: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>