[op][n_upsable][low_end_mode][webos] Limit decoded image bytes in low end mode
:Release Notes:
Avoid pages allocating large image to risk stability.

:Detailed Notes:
For web pages using huge textures as image maps, it is easy to get
to a point where a single image can take 64 or 128MB. That is not
affordable in devices with constrained memory.

In webOS low end mode, set a limit for decoded images that is 1/16
of the available memory for Linux processes. I.e. for 384MB that is
6M pixels. For 640MB, 12.

:Testing Performed:
1. Launch Image gallery
2. Check photos are still OK.

:QA Notes:
Test-cases: QWTC-14546

:Issues Addressed:
[WRQ-3650] Support base::SysInfo::AmountOfPhysicalMemoryMB() and IsLowEndDevice

Change-Id: I06a170dc38b8ebddb074a451f9a9d23829340daf
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/414187
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/419080
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>