[op][upsable][low_end_mode] Allow to force low end mode without faking physical memory
:Release Notes:
Better tuning of memory for low end devices.

:Detailed Notes:
Chromium already provides --enable-low-end-device-mode, that forces
the low end mode, but also fakes the physical memory available to be
512MB. This patch adds a new tune, --force-low-end-device-mode, that
forces low end mode, but without faking the physical memory. We do
this instead of modifying the existing switch as that is used for
testing.

:Testing Performed:
Try with --v=1 launching Chrome with and without the setting. The
flag is enabled when requested.

:QA Notes:
Test-cases: QWTC-14545

:Issues Addressed:
[WRQ-3650] Support base::SysInfo::AmountOfPhysicalMemoryMB() and IsLowEndDevice

Change-Id: If630f038524bede84f8a48b289a25afa6f0017c1
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/414213
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/419085
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>