[op][n_upsable][file_access] Allow blocking access for file access checks
:Release Notes:
Fix DCHECK warnings while accessing local files.

:Detailed Notes:
For local web applications, we implemented checks for file access
so they comply with the web application file access restrictions.
Two of those checks require actually checking if the file exists, and
those calls can be blocking.

Allow them. Ideally this should be replaced by an implementation that
moves the check to a thread worker that allows blocking tasks.

:Testing Performed:
Build with DCHECK and check if TLS assertions are showing.

:QA Issues:
Test-cases: N/A

:Issues Addressed:
[WRQ-12785] webruntime crashes in runtime when built with dcheck_always_on=true
[WRQ-14713] Setup repo and build (m120)

Change-Id: I632420394c8ee85a4e94d4ca6201f5c44e8455b2
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/397115
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/404629
Tested-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>