[op][upsable][low_end_mode][gpu] Allow to tune the number of saved frames
:Release Notes:
Fine tuning of graphics memory usage.

:Detailed Notes:
Add a new switch, --max-numbed-of-saved-frames, that adjust the
limit used in the frame eviction manager. In webOS we already cull
aggressively the backgrounded apps, but for the visible application
our limit will be 5 frames (and 3 in low memory pressure, and 1 in
critical memory pressure).

But we know that in Android, the limit is set to 1 always for memory
lower than 3.5GB. So this switch allows to customize the value for
low end devices.

:Testing Performed:
Check with --v=1 the line Max number of saved frames:

:QA Notes:
Test-cases: QWTC-14542

:Issues Addressed:
[WRQ-3650] Support base::SysInfo::AmountOfPhysicalMemoryMB() and IsLowEndDevice

Change-Id: I3ba623d294d613ce4b8793c5d2f49d3b7a2c3fa2
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/414220
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/419083
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>