[op][n_upsable][luna] Fix LunaServiceClient for permission errors
:Release Notes:
Fix Luna Service permission errors for OEM applications

:Detailed Notes:
The role and client-permission JSON files for OEM applications are
auto generated. The allowed name used in these applications are of
the format <app-id or name>-*
But in LunaServiceClient we have used the format <app-id or name>.*
As a result, the registration to Luna Service fails as permission
error.
So, we have changed "." to "-" and removoed unnecessary webruntime
from camera and audio service names.
Additionally we have added "-" to RegisterApplicationService as well,
since this is going to be used for registering the app for accessing
media controller service

:Testing Performed:
Test audio/camera access:
1. Launch webrtc test application
2. Load https://webcammictest.com/
3. Ensure that mic capture and camera preview are working fine

:QA Notes:
Test-cases: QWTC-6586, QWTC-11390, QWTC-13070
Depends-on: http://gpro.lge.com/c/webosose/wam/+/405012

:Issues Addressed:
[WRQ-18384] Fix Luna Service Permission error for OEM applications

Change-Id: I83543d6a6772a5f76ff159a6e222ed97bc03ca91
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/403820
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/405014
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>