[oe][n_upsable][pwa] Installing and updating PWA using the appinstallservice
:Release Notes:
Implementing installation and update with appinstallservice

:Detailed Notes:
The standard appinstallservice command is used
for installation and upgrades. The luna-command is as follows:
luna://com.webos.appInstallService/install
{ "id": "appId", "ipkUrl": "path_app_data"}
Where path_app_data is the path to the data not packaged
in ipk, and it should start with "pwa://".
The update happens when the application is closed.
Luna-client is implemented as a singleton to provide the correct
procedure for updating application data with appInstallService.

:Testing Performed:
1. Navigate in your browser to a PWA supported site.
2. Click the install icon in the toolbar.
EXP: the icon in the browser disappears.
3. Make sure the installed application is in the launcher.
4. Run this PWA as the ordinary webOS application
EXP: The application should start.
5. Close it using the webOS standard features.
6. Update the webapp in hosting website
7. Run this PWA as the ordinary webOS application
EXP: app is launched
8. Switch to Home
EXP: app icon is updated
9. Uninstall it using the webOS standard features.
10. Verify that the application is removed from the launcher

Websites were used for testing
github.com
bmw.com
tmaiadev-tictactoe.netlify.app
The installation occurred with the following id's
pwa.com.bmw
pwa.com.github
pwa.app.netlify.netlify.tmaiadev-tictactoe
The application is stopped and uninstalled from HOME panel

:QA Notes:
Test-cases: QWTC-13576, QWTC-13585

:Issues Addressed:
[WRQ-14930] Migrage PWA functionality to chromium v.120

Change-Id: I18694f1eb91069290c04aad5cdb4c89ca172af67
Originally-Reviewed-on:
http://gpro.lge.com/c/webosose/chromium108/+/380739
http://gpro.lge.com/c/webosose/chromium108/+/380789
http://gpro.lge.com/c/webosose/chromium108/+/384609
http://gpro.lge.com/c/webosose/chromium108/+/391310
http://gpro.lge.com/c/webosose/chromium108/+/392920
http://gpro.lge.com/c/webosose/chromium108/+/399754
http://gpro.lge.com/c/webosose/chromium108/+/398671
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/405427
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>