[op][n_upsable][browser_shell] Enable 'WebOSServiceBridge' for Luna API call support
:Release Notes:
Fix issue with bookmarks not add to home screen

:Detailed Notes:
This patch is native part of this:
http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/400689
In this, we enable the call to luna API addLaunchPoint of service
com.webos.service.applicationmanager through WebOSServiceBridge.

:Testing Performed:
1. Run the Enact browser application.
2. Navigate to a page that isn't bookmarked.
For example: https://google.com
3. Click on the bookmark button and confirm yes on the shown dialog.
Expectation: A launch point added to home app with title of the page.

:QA Notes:
Test-cases: QWTC-6459

:Issues Addressed:
[WRQ-19026] Enable 'WebOSServiceBridge' in the main page for
            Luna API call support

Change-Id: I8119994a49e1ee54fc9f80c02864e67cba481fc1
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/406218
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>