fixup! [op][n_upsable][browser_shell] Browser Shell crashed in net::SQLitePersistentCookieStore
:Release Notes:
Browser Shell crashed in net::SQLitePersistentCookieStore when
closing the browser by X button

:Detailed Notes:
This crash occurred when calling |DoCommit| in
|net::SQLitePersistentCookieStore::Backend|.It happened because
a darling pointer of |crypto_|.
In this patch, we use neva implementation for CookieCryptoDelegate
which is scoped_refptr to avoid the problem.

:Testing Performed:
Precondition:
Making SSH to raspery device:
1. Run Enact Browser Application by the command:
/usr/bin/browser-shell/run_browser_shell --apppath "file:///usr/palm/
applications/com.webos.app.enactbrowser/index.html"
--appid com.webos.app.enactbrowser
2. Open the Youtube website
3. Close this browser by press to "X" button when this page is loading
The browser should close immediately without delay.
In case crashes continue to happen, get a stacktrace and check that
there are no more information related to the WebRequestAPI function

:QA Notes:
N/A

:Issues Addressed:
[WRQ-22409] Browser Shell crashed in net::SQLitePersistentCookieStore

Change-Id: Ibdd7ae6497bcf34da04c7edea4f1d86d7f06487b
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/408518
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>