[op][ups][cookies] Fix dangling pointer in SQLitePersistentCookieStore::Backend.
:Release Notes:
This commit is to apply upstream CL to fix crash related dangling
pointer in SQLitePersistentCookieStore::Backend. Please see detailed
notes section (from Chromium) for further details.

:Detailed Notes:
The crypto_ pointer member was raw and it was possible for
this to be invalidated and accessed. This could happen because
the lifetime of the SQLite backend is refcounted and self-owned
and it might outlive network context and network service which
can both be destroyed before the Backend.

One possible scenario here is if the DoCommit() takes some time
and could still be running on the Backend background sequence
and try to access the crypto_ object to perform encryption to
write data.

This CL makes two substantive changes:

1. The pattern of ownership for the cookie_crypto_delegate is no
longer a raw pointer whose lifetime must be managed by the
caller, but is replaced by a unique_ptr strongly owned by
the Backend, created in the Network Context when required.

2. The interface to the cookie encryption provider is changed
to be specific to each network context, and not configured on
the network service, similar to existing cookie encryption. This
means there to be a single unique_ptr for the delegate per
network context and encryption configuration can be controlled
on a context-by-context basis.

The unique_ptr is now passed into the SQLitePersistentCookieStore
and lifetime is thus ensured. Several related interfaces are
updated to use this unique_ptr pattern, however, this CL does not
contain any functional changes.

A test interface is added to reproduce this failure by adding
a delay on commits, and the existing tests that use OSCrypt Async
are augmented to add this delay. The test previously crashed before
this CL.

BUG=1511730,1426805

:Testing Performed:
Build only.

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRQ-22906] Apply upstream CL related to crash in
            net::SQLitePersistentCookieStore

Change-Id: Idc044af001c7e81e0c23fc12b008ec7b7837b530
Originally-Reviewed-on:
https://chromium-review.googlesource.com/c/chromium/src/+/5121037
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/413757
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>