[oe][n_upsable][pwa] Rework of opening external links in the browser
:Release Notes:
It was necessary to develop navigation inside iframes and navigation
on friendly sites

:Detailed Notes:
When a PWA application follows a link that has an origin other than PWA,
it must be opened in the web browser of that OS.
Not all link navigations are external, we need to take into account
transitions inside iframes and transitions to friendly sites.
For this purpose, the algorithm of external links verification
was finalized. The "Content-Security-Policy" from ResponseHeaders
is analyzed. From the fields "frame-src, default-src, script-src"
a regular expression is formed for comparison with the url for the
transition. The presence of RenderFrameHost during navigation is also
taken into account.

:Testing Performed:
1. Open PWA app
2. Click external link (to page outside of the PWA site)

EXP: no navigation on the PWA page, enactbrowser is launched and
navigated to the link

:QA Notes:
Test-cases: QWTC-13576, QWTC-13585, QWTC-13630

:Issues Addressed:
[WRQ-27435] Provide a solution for composing a list of internal
site URLs

Change-Id: I788b07760f9b2a780fd495196555c93153b7fd01
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/420063
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>