[op][n_upsable][browser_shell] Fix crash when visiting messenger.com
:Release Notes:
Fix browser shell crash when the application checks can access
data for origin for Facebook Messenger.

:Detailed Notes:
When loading a site request an isolation process:
1. In chromium108:
The method
In handle to check, assign and lock process to site.
It's just return true in method CanAccessDataForMaybeOpaqueOrigin
by adding #if !BUILDFLAG(IS_ANDROID). Then skip handle
for request an isolation process here.
And then the task for handle isolation process will handle
in render host process.

2. In chromium120:
When have an task a site request an isolation process.
Browser-shell(chromium120) have handle in both in
content/browser and content/browser/render_host.
Lead handle in content/browser
can not pass CHECK policy called.
By default, it should be handled only in the render host process.
So in this patch,we make a workaround solution
(disable by default for kSiteIsolationCitadelEnforcement
to make the handle same adding
#if !BUILDFLAG(IS_ANDROID) as chromium108 did)

:Testing Performed:
1. Visit https://www.messenger.com or https://www.facebook.com
2. Login with your facebook account
3. Expect the message.com page is loaded and
the content is shown properly

:QA Notes:
Test-cases: QWTC-14810

:Issues Addressed:
[WRQ-31079] The Enact Browser crashed when visiting messenger.com

Change-Id: Iaa830f0591165c55ccbfe71b56a2e52b73f8b0c3
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/426486
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>