[op][n_upsable][browser_shell] Fix crash when press back and forward button
:Release Notes:
Fix crash when press back and forward button

:Detailed Notes:
When users quickly press the back and forward buttons multiple times,
the canGoBack and canGoForward states that are passed to the JavaScript
site do not update correctly, which leads to a crash. To deal with
this problem, I will check every time the GoBack and GoForward events
are triggered.

:Testing Performed:
1. Open the Enact browser application.
2. Navigate to the Google website.
3. Input "youtube" in the Google search field and press the Enter key.
4. This will navigate to the YouTube website from the Google search
results.
5. Quickly press the back button 3 times.
6. Quickly press the forward button 3 times.
Expected: The crash mentioned in steps 5 and 6 should not happen.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-28574] [RDX_CRASH][webos] /usr/bin/browser-shell/browser_shell
            in n/a (libcbe.so + 0x9b7e284)

Change-Id: I476f1682c947092891bdedc6970be76078053f91
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/426560
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Dat Pham <dat2.pham@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Jeongran Jang <jeongran.jang@lge.com>