[op][n_upsable][browser_shell] Fix crash when closing browser after trying to download a PDF
:Release Notes:
Fix issue: Crash when closing browser after trying to download a PDF

:Detailed Notes:
Whenever trying to downnload a PDF, an object of neva_app_runtime
::AppRuntimeDownloadManagerDelegate will be constructed, and it is
a member of AppRuntimeBrowserContext. When closing browser,
browser context in neva_app_runtime is destructed and its member
is destructed also. Then in DownloadManagerImpl::Shutdown(), because
its delegate_ was deleted before, a crash happens.

:Testing Performed:
1. Open browser and open a PDF file.
2. Click "Download" button (ignore warning or message)
3. Close browser by clicking "X" button
4. Expected result: browser closed normally, and no core dumps
exist in journal log.

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRQ-30139] [RDX_CRASH][webos] /usr/bin/browser-shell/browser_
shell in __close (libc.so.6 + 0xdb2f4)

Change-Id: I905f9f22448920778f804f74a6a6cf0921ee1c83
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/426713
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Nhan Do <nhan.do@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jeongran Jang <jeongran.jang@lge.com>