[op][ups][icu] Fix null termination
:Release Notes:
Fix null termination of ICU that makes broken timezone strings.

:Detailed Notes:
When timezone is changed, the timezone string is not properly set
when the new value is shorter than the previous value.
For example, when timezone is changed from "America/New_York" to
"America/Adak", the string is set to "America/AdakYork".
Because of broken strings, NotifyTimezoneChangeToV8() is not called,
and new Date() does not use the new timezone.

There was a bug of ICU related to the null termination of strings.
The related upstream bug is https://crbug.com/41495940.
It was fixed by https://crrev.com/c/5299470.

:Testing Performed:
Change timezone and check the time is properly updated.

:QA Notes:
Test-cases: QWTC-6094, QWTC-9393

:Issues Addressed:
[WRQ-28865] [OSE -RPI4]: Time and Date is not changed with respective to
            Time zone
[WRQ-31472] [TAS Failed][OSE][QWTC-9393] [Settings] Verify Time is
            updated while Time Zone is changed

Change-Id: I11e4ea4c1865d65554f24c0815a58de77e0575cd
Originally-Reviewed-on: https://crrev.com/c/5299470
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/427233
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Blink-Build: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Tested-by: Blink Jenkins <blink.jenkins@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>