fixup! [oe][n_upsable][audio] Fix pulse audio input/output device id
:Release Notes:
Fix default and preferred device id for WebOSAudioInputStream

:Detailed Notes:
In case when user does not select any device from the list we get
"default" as the device name. But we have to use "webcall" as the
default device id for pulse audio input stream and the preferred
device should be empty.
But with the changes done for handling dual display case "default"
was being set as the preferred device.

So, in this patch set we check both "default" and "default with
display id" as the default device making preferred device is set
only when it is not default.

:Testing Performed:
1. Install test.redirection.webosvda App
2. launch app with https://mictests.com/
3. Click "Test on mic"
4. Speak something on mic
5. click play on the audio generated
6. Recordered mic input should be heard on the speaker

:QA Notes:
Test-cases: QWTC-11387

:Issues Addressed:
[WRQ-30331] [OSE -RPI4]: Microphone device is not detected and
            audio is not heard.

Change-Id: Id68da905f836d97d5d8ec95c5a5cbc962809fe95
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/427455
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Tested-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>