[op][n_upsable][browser_shell] Post pending messages after the injection IPC object is ready
:Release Notes:
Improved injection IPC logic to ensure pending messages are posted only
after the IPC object is ready

:Detailed Notes:
Currently, our injection IPC mechanism is implemented asynchronously
among the browser process, renderer process, and JS V8.
Whenever an IPC object in JS is created, it maps with the object's
lifecycle and the connection flow is as follows:
- [V8] IPC JS object is created.
- [Renderer Process] IPC object BrowserShellIpcInjection is created.
- [Renderer Process] Sends a bind request to create a connection to
  the Browser process IPC object (ShellIpc).
- [Browser Process] ShellIpc object is created.
- [Renderer Process] Setup callback is called.
- [Renderer Process] Bind remote connection <-> Pipe established.

However, because the UIOverlay logic from JS requires two sequential
instructions:
- Create a new IPC JS object.
- Post a "ready" message to the receiver to create the pop-up.
The "ready" message is still executed by
BrowserShellIpcInjection::Post, but it is stuck at
ShellIpcEndpointImpl::Handle because the connection readiness is
checked (is_bound()). As a result, the message is not sent back to
the receiver on the JS side, which is why the UIOverlay is not shown
on the JS side.

:Testing Performed:
1. Visit https://www.google.com
2. Click/short-touch on [Menu] or Zoom button on the top-right
in Menu frame for multiple times
3. Click/short-touch on List Extension button for multiple times
4. Click/short-touch on Zoom/Exit Zoom button for multiple times
5. Click/short-touch on Bookmark button for multiple times
6. Visit testpages.herokuapp.com/styled/alerts/alert-test.html
7. Click all button in this page for multiple times

:QA Notes:
Test-cases: QWTC-11155

:Issues Addressed:
[WRQ-31629] The menu dropdown list does not display when clicking
            the menu button

Change-Id: I1141db9bdd3b9ec8cd71b738904d4134f970bbf3
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/429078
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Blink Jenkins <blink.jenkins@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>