[op][n_upsable][browser_shell] Fix crash possibly occurring in permission prompt logic
:Release Notes:
Fix a crash that can occur when a permission prompt is displayed to
request permission (e.g., for microphone, audio, geolocation, etc.) due
to a strong binding of an object using base::Unretained.

:Detailed Notes:
The current implementation of the permission prompt logic uses
base::Unretained to bind a callback with the permission delegate.
However, this object |PermissionPromptWebOS| can be destroyed before
the callback is called, leading to a crash in the browser process and
causing the entire browser application to crash.
To resolve this issue, we need to bind the callback to a weak object.

Please see details below. We can see that permission prompt is
triggered by |WebContents| -> |PermissionRequestManager|.
|PermissionRequestManager| owns |view_|, |view_| is
|PermissionPromptWebOS| and created by
|NevaPermissionsClient::CreatePrompt|, it owns an ivar
|BrowserShellPermissionPromptDelegate|.

When a prompt is shown it means that an |PermissionPromptWebOS|
created by calling
|BrowserShellPermissionsClientDelegate::CreatePrompt|.
Then BrowserShellPermissionPromptDelegate::ShowBubble called in its
constructor |PermissionPromptWebOS| and then we have a callback
(BrowserShellPermissionPromptDelegate::OnPromptResponse bound with
the delegate object |BrowserShellPermissionPromptDelegate| by
base::Unretained) passed.

|view_| can be destroyed by some contexts (WebContents deleted,
visibility changed, etc.), however, an action (press ok, cancel,
etc.) can triggered after that which make callback called and then
crash occurred.

:Testing Performed:
Build only.

:QA Notes:
Test-cases: QWTC-10173,QWTC-10174

:Issues Addressed:
[WRQ-33769] [RDX_CRASH][webos] /usr/bin/browser-shell/browser_shell
            in clock_nanosleep (libc.so.6 + 0xb4dd0)

Change-Id: I5b7879aed6701d8dcc23b57ea8f81680bf5ff9ae
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/431396
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>