[op][n_ups][appruntime] Add shutdown logic for KeyedService of the BrowserContext
:Release Notes:
Added shutdown logic for KeyedService of the BrowserContext

:Detailed Notes:
The shutdown logic for KeyedService of the BrowserContext was missing
for the AppRuntimeBrowserContext.
So the KeyedService like NevaExtensionsServiceImpl was not properly
terminated on the shutdown.
This CL adds the shutdown logic which terminate all KeyedService that
is related to the given BrowserContext.

:Testing Performed:
Close the Neva browser and verify that there is no crash.

:QA Notes:

:Issues Addressed:
[WRR-2265] [Runtime Memory] Port NEVA patches regarding "Using
extensions only in contexts for third-party content"

Change-Id: Ibcc0a5cb36a4a9bf1a14e4737955ca25202bd57a
Originally-reviewed-on: http://mercury.lge.net:8080/c/neva/chromium/+/10438
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/434959
Blink-Build: Nguyen Huy Quang <quang14.nguyen@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Blink Jenkins <blink.jenkins@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Tested-by: Nguyen Huy Quang <quang14.nguyen@lge.com>