fixup! [appruntime][extensions][injection] Implement ExtensionsServicesManager
:Release Notes:
Fixed not to use std::make_unique when creating NevaExtensionsServiceImpl object

:Detailed Notes:
This CL fixes not to use std::make_unique when creating the
NevaExtensionsServiceImpl object which was introduced on:
http://mercury.lge.net:8080/c/neva/chromium/+/10842

The NevaExtensionsServiceImpl is a keyed service and the key is the
BrowserContext.
However, all the keyed services's destruction for a specific
BrowserContext is managed by the DependencyManager class.
So the keyed services which the key is the BrowserContext should not
be created as a smart pointer.

:Testing Performed:
1. Run Neva browser with MetaMask extension. (See NEVA-8703 for details)
2. Close the Neva browser and verify that there is no crash.

:QA Notes:

:Issues Addressed:
[WRR-2265] [Runtime Memory] Port NEVA patches regarding "Using
extensions only in contexts for third-party content"
[NEVA-8759] If I press the X button to quit after loading a random
page in Enactbrowser, an error occurs
[NEVA-8910] Revise BrowserContext usage in the BrowserShell

Change-Id: I82c6c63086fec1b713f95e82fa7de1158e4c7729
Originally-reviewed-on: http://mercury.lge.net:8080/c/neva/chromium/+/11164
Reviewed-on: http://gpro.lge.com/c/webosose/chromium120/+/434960
Tested-by: Blink Jenkins <blink.jenkins@lge.com>
Blink-Build: Nguyen Huy Quang <quang14.nguyen@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>